relocate store
